Projektas

+++ app/Http/Controllers/AdminController.php
@@ -5,28 +5,36 @@ namespace App\Http\Controllers;
 use App\Models\Game;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 
 class AdminController extends Controller
 {
     public function listAllGames()
     {
-        $games = Game::all();
-        return Inertia::render('Admin/AllGames', ['games' => $games]);
+        if (Auth::check()) {
+            $games = Game::all();
+            return Inertia::render('Admin/AllGames', ['games' => $games]);
+        }
+        redirect('/')->with('error', "You're not an administrator!");
     }
 
     public function destroyGame(int $id, Game $game)
     {
-        $game = Game::findOrFail($id);
-        $game->delete();
-        return redirect('/')->with('success', 'Game removed!');
+        if (Auth::check()) {
+            $game = Game::findOrFail($id);
+            $game->delete();
+            return redirect('/')->with('success', 'Game removed!');
+        }
+        redirect('/')->with('error', "You're not an administrator!");
     }
 
-    public function listAllUsers(){
-        $users = User::all();
-        return Inertia::render('Admin/AllUsers', ['users' => $users]);
-
-
-        
+    public function listAllUsers()
+    {
+        if (Auth::check()) {
+            $users = User::all();
+            return Inertia::render('Admin/AllUsers', ['users' => $users]);
+        }
+        redirect('/')->with('error', "You're not an administrator!");
     }
 }
